requirements.txt
Flask==2.1.2
Flask-JWT-Extended==4.4.2
flask-marshmallow==0.14.0
Flask-Pydantic==0.9.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
httptools==0.4.0
psycopg2-binary==2.9.3
pytest==7.1.2
flasgger==0.9.5
gevent==21.12.0
redis==4.3.4
python-dotenv==0.19.2
cryptography==37.0.4
certifi==2022.5.18.1
cffi==1.15.1
bcrypt==3.2.2
alembic==1.8.1
rauth==0.7.3

